Inline component we don't use elsewhere.
